Rename dprint to console.rs to avoid the issue where dprint module has a macro called dprint which makes the use statement seem odd.
